moved all cron jobs to special service tasks.service. moved all payments to special service payments.service. added all timezones
