+++ src/server.js
@@ -5,7 +5,7 @@ const cors = require('cors')
 const cookieParser = require('cookie-parser')
 const router = require('./routes')
 const errorMiddleware = require('./middlewares/error.middleware')
-const cartService = require('./services/domain/cart.service')
+const tasksService = require('./services/domain/tasks.service')
 
 const app = express()
 
@@ -14,7 +14,7 @@ app.use(cookieParser())
 app.use(cors())
 app.use('/api/v1', router)
 app.use(errorMiddleware)
-cartService.scheduleReleaseOfOldReservations()
+tasksService.start()
 
 const PORT = process.env.PORT ?? 8080
